ADD: resources to build and deploy python package
