+++ setup.cfg
@@ -1,0 +1,28 @@
+[metadata]
+name = youssemd_diffusion2d
+version = 0.0.1
+author = Mohamed Youssef
+author_email = mohamedloutfy97@gmail.com
+description = A Python package for solving the 2D diffusion equation using Finite Difference Method
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/MohamedAlyLoutfy/diffusion2D.git
+classifiers =
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3.6
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
+    License :: OSI Approved :: MIT License
+    Operating System :: OS Independent
+
+[options]
+packages = find:
+python_requires = >=3.6
+install_requires =
+    numpy
+    matplotlib
+
+[options.package_data]
+* = *.txt, *.md
