+++ youssemd_diffusion2d.egg-info/SOURCES.txt
@@ -1,0 +1,12 @@
+LICENSE
+README.md
+setup.cfg
+setup.py
+youssemd_diffusion2d/__init__.py
+youssemd_diffusion2d/diffusion2d.py
+youssemd_diffusion2d/output.py
+youssemd_diffusion2d.egg-info/PKG-INFO
+youssemd_diffusion2d.egg-info/SOURCES.txt
+youssemd_diffusion2d.egg-info/dependency_links.txt
+youssemd_diffusion2d.egg-info/requires.txt
+youssemd_diffusion2d.egg-info/top_level.txt
